Restore websocket message for updating tweets
Did a much better job of it this time around. Now we have specific
protobuf messages for different types of events, and they are all sent
to an events channel. The Elm side decodes them into a union type, and
then each can be handled appropriately.

I think this will scale well to other sorts of liveness that we need to
add to the app.