Pull out common page functionality
This a huge refactor. The goal here is to make sure common things many
pages (maybe not all) need to do isn't repeated for every page. It means
less friction to add a modal or errors or whatever to a page that
doesn't already have them.

Also fixed showing the navbar menu on mobile, which was the change that
prompted this giant refactor.