Complete Elm 0.19 upgrade
This could not really be done incrementally.

The biggest change was removing the ActionCable package and changing to
handling events with ports, since the Elm websocket package has not been
updated for Elm 0.19. Honestly, the result is so much simpler, since we
rely on the native JS ActionCable library and our needs are really
simple. I wish I had done it this way from the beginning.