Rework account screen
Fix subscribe button, which broke in the Elm 0.19 upgrade. There's no
longer a way to embed a script tag in Elm HTML, so we needed to switch
to the custom JS way of using Stripe Checkout. This ends up being a more
Elm-y solution anyway, and we don't need to have a non-Twirp endpoint
for it, so that's nice.